Get checksums from original tex.xbx file.
ug2-dds2tex now accepts the original tex.xbx file as input using the -c
switch. Checksums(?) will be copied over from this to the new version of
the file. This is necessary because THUG2/THUG Pro will not display the textures
properly if this value does not match the original. It doesn't seem to be a CRC32
of any part of the file, so it's probably a reference id hardcoded into
a level's scripts.