Bugfix: InflateFile was writing invalid memory.
When reading the final type byte, I wasn't checking if there were
actually enough bytes to process. In the process of tracking down
the problem, I've moved to using globally accessible vectors that
I resize once to the size of the largest file that will be copied
to/from them.